tfi plugins possible
- Creation of all things to enable plugins easily :
  - A new options for plugins in admin_panel called tfi_plugins
  - This option returns if a plugin is used or not
  - PluginsManager is a class which knows all plugins
  - If there is a new plugin, it will be added to the plugin list
  - When the option is updated, plugins are activate or deactivate
  - Multiple hooks for activation/deactivation
    - tfi_plugins_activate_{plugin_name} for activation
    - tfi_plugins_deactivate_{plugin_name} for deactivation
  - Add a uninstall.php file on the plugin directory to call it when
    we want to uninstall the plugin

+++ utilities/functions.php
@@ -114,7 +114,13 @@ if ( ! function_exists( 'tfi_delete_files' ) ) {
                 tfi_delete_files( $file );      
             }
 
-            rmdir( $target );
+            /**
+             * Verification of the file, because if the '/' isn't set at the end of the first target,
+             * the folder is also inside $files and will be deleted twice, generating a warning
+             */
+            if ( file_exists( $target ) ) {
+                rmdir( $target );
+            }
         }
         else if ( is_file( $target ) ) {
             unlink( $target );
